skill/requirements.txt
functions-framework==3.*
google-cloud-logging
box-sdk-gen[jwt]